Renamed to jsx files and rearranged frontend files.
